Extracted interface from PropertyTypeResolver in case people want more elaborate replacements
